separate airdrop tasks
